Frontend: - Extracted CommentThreadItem into separate file.
